requirements.txt
# Agentic RAG Temel Kütüphaneleri (Uyumlu sürümler için güncellenmiştir)
openai>=1.109.1,<3.0.0
langchain
langchain-openai
python-dotenv
pydantic
loguru
tiktoken
simpleeval
PyYAML
PyPDF2
pymupdf